Migrated confirmation page to SCSS modules.
