+++ src/components/confirmation/ConfirmationDialog/index.module.scss
@@ -1,0 +1,63 @@
+@import "~/styles/variables.scss";
+@import "~/styles/mixins.scss";
+
+.dialog {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 50%;
+  border-radius: 25px; 
+  padding: 2em;
+  margin: 0 auto;
+  flex-direction: column;
+  background-image: linear-gradient(316.54deg, rgba(255, 227, 227, 0.7565) 
+  35.13%, rgba(255, 255, 255, 0.85) 126.39%);
+  box-shadow: 0px 4px 4px rgba(200, 116, 56, 0.25);
+  backdrop-filter: blur(4px); 
+}
+
+.dialogContent {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.dialogText {
+  margin-bottom: 1em;
+  color: $wave-gradient-color-start; 
+}
+
+.headerContainer {
+  text-align: "center";
+  align-self: "center";
+  
+  @include down($breakpoint-md){
+    width: 60%; 
+  }
+
+  @include down($breakpoint-sm){
+    width: 80%;
+  }
+
+  padding-bottom: 0.5rem;
+}
+
+.header {
+  border-bottom: solid $text-color-primary 2px;
+  padding-bottom: 0.5rem;
+
+  @include down($breakpoint-md){
+    fontSize: 2.5em; 
+  }
+
+  @include donw($breakpoint-sm){
+    fontSize: 1.8em; 
+  }
+
+  color: $text-color-primary; 
+  font-weight: 600; 
+}
+
+.link {
+  color: $text-color-primary; 
+}
