+++ src/components/design/ScottyLabsHeader/index.tsx
@@ -1,28 +1,12 @@
 import React, { ReactElement } from "react"
-import { Link, makeStyles, Theme } from "@material-ui/core"
+import {Link} from "@material-ui/core"
 import NextLink from "next/link"
-import TartanHacksIcon from "./TartanHacksIcon"
-import { useTheme } from "@material-ui/styles"
-
-const useStyles = makeStyles((theme) => ({
-  pageHeader: {
-    position: "absolute",
-    left: "3%",
-    top: "4%",
-    whiteSpace: "nowrap",
-    width: "45%",
-    [theme.breakpoints.down(theme.breakpoints.values.mobile)]: {
-      width: "80%"
-    }
-  }
-}))
+import TartanHacksIcon from "../TartanHacksIcon"
+import styles from "./index.module.scss"
 
 const ScottyLabsHeader = (): ReactElement => {
-  const classes = useStyles()
-  const theme: Theme = useTheme()
-
   return (
-    <div className={classes.pageHeader}>
+    <div className={styles.pageHeader}>
       <NextLink href="/" passHref>
         <Link>
           <svg
@@ -33,15 +17,15 @@ const ScottyLabsHeader = (): ReactElement => {
           >
             <path
               d="M120.112 6.496V11.92H111.184V40H104.464V11.92H95.536V6.496H120.112ZM123.287 26.608C123.287 23.92 123.815 21.536 124.871 19.456C125.959 17.376 127.415 15.776 129.239 14.656C131.095 13.536 133.159 12.976 135.431 12.976C137.415 12.976 139.143 13.376 140.615 14.176C142.119 14.976 143.319 15.984 144.215 17.2V13.408H150.983V40H144.215V36.112C143.351 37.36 142.151 38.4 140.615 39.232C139.111 40.032 137.367 40.432 135.383 40.432C133.143 40.432 131.095 39.856 129.239 38.704C127.415 37.552 125.959 35.936 124.871 33.856C123.815 31.744 123.287 29.328 123.287 26.608ZM144.215 26.704C144.215 25.072 143.895 23.68 143.255 22.528C142.615 21.344 141.751 20.448 140.663 19.84C139.575 19.2 138.407 18.88 137.159 18.88C135.911 18.88 134.759 19.184 133.703 19.792C132.647 20.4 131.783 21.296 131.111 22.48C130.471 23.632 130.151 25.008 130.151 26.608C130.151 28.208 130.471 29.616 131.111 30.832C131.783 32.016 132.647 32.928 133.703 33.568C134.791 34.208 135.943 34.528 137.159 34.528C138.407 34.528 139.575 34.224 140.663 33.616C141.751 32.976 142.615 32.08 143.255 30.928C143.895 29.744 144.215 28.336 144.215 26.704ZM164.266 17.536C165.13 16.128 166.25 15.024 167.626 14.224C169.034 13.424 170.634 13.024 172.426 13.024V20.08H170.65C168.538 20.08 166.938 20.576 165.85 21.568C164.794 22.56 164.266 24.288 164.266 26.752V40H157.546V13.408H164.266V17.536ZM184.777 18.928V31.792C184.777 32.688 184.985 33.344 185.401 33.76C185.849 34.144 186.585 34.336 187.609 34.336H190.729V40H186.505C180.841 40 178.009 37.248 178.009 31.744V18.928H174.841V13.408H178.009V6.832H184.777V13.408H190.729V18.928H184.777ZM193.834 26.608C193.834 23.92 194.362 21.536 195.418 19.456C196.506 17.376 197.962 15.776 199.786 14.656C201.642 13.536 203.706 12.976 205.978 12.976C207.962 12.976 209.69 13.376 211.162 14.176C212.666 14.976 213.866 15.984 214.762 17.2V13.408H221.53V40H214.762V36.112C213.898 37.36 212.698 38.4 211.162 39.232C209.658 40.032 207.914 40.432 205.93 40.432C203.69 40.432 201.642 39.856 199.786 38.704C197.962 37.552 196.506 35.936 195.418 33.856C194.362 31.744 193.834 29.328 193.834 26.608ZM214.762 26.704C214.762 25.072 214.442 23.68 213.802 22.528C213.162 21.344 212.298 20.448 211.21 19.84C210.122 19.2 208.954 18.88 207.706 18.88C206.458 18.88 205.306 19.184 204.25 19.792C203.194 20.4 202.33 21.296 201.658 22.48C201.018 23.632 200.698 25.008 200.698 26.608C200.698 28.208 201.018 29.616 201.658 30.832C202.33 32.016 203.194 32.928 204.25 33.568C205.338 34.208 206.49 34.528 207.706 34.528C208.954 34.528 210.122 34.224 211.21 33.616C212.298 32.976 213.162 32.08 213.802 30.928C214.442 29.744 214.762 28.336 214.762 26.704ZM242.829 13.024C245.997 13.024 248.557 14.032 250.509 16.048C252.461 18.032 253.437 20.816 253.437 24.4V40H246.717V25.312C246.717 23.2 246.189 21.584 245.133 20.464C244.077 19.312 242.637 18.736 240.813 18.736C238.957 18.736 237.485 19.312 236.397 20.464C235.341 21.584 234.813 23.2 234.813 25.312V40H228.093V13.408H234.813V16.72C235.709 15.568 236.845 14.672 238.221 14.032C239.629 13.36 241.165 13.024 242.829 13.024ZM287.62 6.496V40H280.9V25.744H266.548V40H259.828V6.496H266.548V20.272H280.9V6.496H287.62ZM292.506 26.608C292.506 23.92 293.034 21.536 294.09 19.456C295.178 17.376 296.634 15.776 298.458 14.656C300.314 13.536 302.378 12.976 304.65 12.976C306.634 12.976 308.362 13.376 309.834 14.176C311.338 14.976 312.538 15.984 313.434 17.2V13.408H320.202V40H313.434V36.112C312.57 37.36 311.37 38.4 309.834 39.232C308.33 40.032 306.586 40.432 304.602 40.432C302.362 40.432 300.314 39.856 298.458 38.704C296.634 37.552 295.178 35.936 294.09 33.856C293.034 31.744 292.506 29.328 292.506 26.608ZM313.434 26.704C313.434 25.072 313.114 23.68 312.474 22.528C311.834 21.344 310.97 20.448 309.882 19.84C308.794 19.2 307.626 18.88 306.378 18.88C305.13 18.88 303.978 19.184 302.922 19.792C301.866 20.4 301.002 21.296 300.33 22.48C299.69 23.632 299.37 25.008 299.37 26.608C299.37 28.208 299.69 29.616 300.33 30.832C301.002 32.016 301.866 32.928 302.922 33.568C304.01 34.208 305.162 34.528 306.378 34.528C307.626 34.528 308.794 34.224 309.882 33.616C310.97 32.976 311.834 32.08 312.474 30.928C313.114 29.744 313.434 28.336 313.434 26.704ZM325.037 26.704C325.037 23.952 325.597 21.552 326.717 19.504C327.837 17.424 329.389 15.824 331.373 14.704C333.357 13.552 335.629 12.976 338.189 12.976C341.485 12.976 344.205 13.808 346.349 15.472C348.525 17.104 349.981 19.408 350.717 22.384H343.469C343.085 21.232 342.429 20.336 341.501 19.696C340.605 19.024 339.485 18.688 338.141 18.688C336.221 18.688 334.701 19.392 333.581 20.8C332.461 22.176 331.901 24.144 331.901 26.704C331.901 29.232 332.461 31.2 333.581 32.608C334.701 33.984 336.221 34.672 338.141 34.672C340.861 34.672 342.637 33.456 343.469 31.024H350.717C349.981 33.904 348.525 36.192 346.349 37.888C344.173 39.584 341.453 40.432 338.189 40.432C335.629 40.432 333.357 39.872 331.373 38.752C329.389 37.6 327.837 36 326.717 33.952C325.597 31.872 325.037 29.456 325.037 26.704ZM371.384 40L362.36 28.672V40H355.64V4.48H362.36V24.688L371.288 13.408H380.024L368.312 26.752L380.12 40H371.384ZM393.847 40.432C391.671 40.432 389.719 40.048 387.991 39.28C386.263 38.48 384.887 37.408 383.863 36.064C382.871 34.72 382.327 33.232 382.231 31.6H388.999C389.127 32.624 389.623 33.472 390.487 34.144C391.383 34.816 392.487 35.152 393.799 35.152C395.079 35.152 396.071 34.896 396.775 34.384C397.511 33.872 397.879 33.216 397.879 32.416C397.879 31.552 397.431 30.912 396.535 30.496C395.671 30.048 394.279 29.568 392.359 29.056C390.375 28.576 388.743 28.08 387.463 27.568C386.215 27.056 385.127 26.272 384.199 25.216C383.303 24.16 382.855 22.736 382.855 20.944C382.855 19.472 383.271 18.128 384.103 16.912C384.967 15.696 386.183 14.736 387.751 14.032C389.351 13.328 391.223 12.976 393.367 12.976C396.535 12.976 399.063 13.776 400.951 15.376C402.839 16.944 403.879 19.072 404.071 21.76H397.639C397.543 20.704 397.095 19.872 396.295 19.264C395.527 18.624 394.487 18.304 393.175 18.304C391.959 18.304 391.015 18.528 390.343 18.976C389.703 19.424 389.383 20.048 389.383 20.848C389.383 21.744 389.831 22.432 390.727 22.912C391.623 23.36 393.015 23.824 394.903 24.304C396.823 24.784 398.407 25.28 399.655 25.792C400.903 26.304 401.975 27.104 402.871 28.192C403.799 29.248 404.279 30.656 404.311 32.416C404.311 33.952 403.879 35.328 403.015 36.544C402.183 37.76 400.967 38.72 399.367 39.424C397.799 40.096 395.959 40.432 393.847 40.432Z"
-              fill={theme.palette.text.primary}
+              fill={styles.primaryTextColor}
             />
             <path
               d="M420.772 28.275C421.255 27.6083 421.914 27.0667 422.747 26.65C423.597 26.2167 424.539 26 425.572 26C426.789 26 427.889 26.2917 428.872 26.875C429.855 27.4583 430.63 28.2917 431.197 29.375C431.764 30.4417 432.047 31.6667 432.047 33.05C432.047 34.4333 431.764 35.675 431.197 36.775C430.63 37.8583 429.847 38.7083 428.847 39.325C427.864 39.925 426.772 40.225 425.572 40.225C424.505 40.225 423.555 40.0167 422.722 39.6C421.905 39.1833 421.255 38.65 420.772 38V40H417.922V21.5H420.772V28.275ZM429.147 33.05C429.147 32.1 428.947 31.2833 428.547 30.6C428.164 29.9 427.647 29.375 426.997 29.025C426.364 28.6583 425.68 28.475 424.947 28.475C424.23 28.475 423.547 28.6583 422.897 29.025C422.264 29.3917 421.747 29.925 421.347 30.625C420.964 31.325 420.772 32.15 420.772 33.1C420.772 34.05 420.964 34.8833 421.347 35.6C421.747 36.3 422.264 36.8333 422.897 37.2C423.547 37.5667 424.23 37.75 424.947 37.75C425.68 37.75 426.364 37.5667 426.997 37.2C427.647 36.8167 428.164 36.2667 428.547 35.55C428.947 34.8333 429.147 34 429.147 33.05ZM447.265 26.225L438.815 46.475H435.865L438.665 39.775L433.24 26.225H436.415L440.29 36.725L444.315 26.225H447.265ZM467.837 40.252C466.157 40.252 464.645 39.964 463.301 39.388C461.957 38.788 460.901 37.948 460.133 36.868C459.365 35.788 458.981 34.528 458.981 33.088H463.373C463.469 34.168 463.889 35.056 464.633 35.752C465.401 36.448 466.469 36.796 467.837 36.796C469.253 36.796 470.357 36.46 471.149 35.788C471.941 35.092 472.337 34.204 472.337 33.124C472.337 32.284 472.085 31.6 471.581 31.072C471.101 30.544 470.489 30.136 469.745 29.848C469.025 29.56 468.017 29.248 466.721 28.912C465.089 28.48 463.757 28.048 462.725 27.616C461.717 27.16 460.853 26.464 460.133 25.528C459.413 24.592 459.053 23.344 459.053 21.784C459.053 20.344 459.413 19.084 460.133 18.004C460.853 16.924 461.861 16.096 463.157 15.52C464.453 14.944 465.953 14.656 467.657 14.656C470.081 14.656 472.061 15.268 473.597 16.492C475.157 17.692 476.021 19.348 476.189 21.46H471.653C471.581 20.548 471.149 19.768 470.357 19.12C469.565 18.472 468.521 18.148 467.225 18.148C466.049 18.148 465.089 18.448 464.345 19.048C463.601 19.648 463.229 20.512 463.229 21.64C463.229 22.408 463.457 23.044 463.913 23.548C464.393 24.028 464.993 24.412 465.713 24.7C466.433 24.988 467.417 25.3 468.665 25.636C470.321 26.092 471.665 26.548 472.697 27.004C473.753 27.46 474.641 28.168 475.361 29.128C476.105 30.064 476.477 31.324 476.477 32.908C476.477 34.18 476.129 35.38 475.433 36.508C474.761 37.636 473.765 38.548 472.445 39.244C471.149 39.916 469.613 40.252 467.837 40.252ZM479.951 30.064C479.951 28.024 480.359 26.236 481.175 24.7C482.015 23.14 483.167 21.94 484.631 21.1C486.095 20.26 487.775 19.84 489.671 19.84C492.071 19.84 494.051 20.416 495.611 21.568C497.195 22.696 498.263 24.316 498.815 26.428H494.387C494.027 25.444 493.451 24.676 492.659 24.124C491.867 23.572 490.871 23.296 489.671 23.296C487.991 23.296 486.647 23.896 485.639 25.096C484.655 26.272 484.163 27.928 484.163 30.064C484.163 32.2 484.655 33.868 485.639 35.068C486.647 36.268 487.991 36.868 489.671 36.868C492.047 36.868 493.619 35.824 494.387 33.736H498.815C498.239 35.752 497.159 37.36 495.575 38.56C493.991 39.736 492.023 40.324 489.671 40.324C487.775 40.324 486.095 39.904 484.631 39.064C483.167 38.2 482.015 37 481.175 35.464C480.359 33.904 479.951 32.104 479.951 30.064ZM511.509 40.324C509.637 40.324 507.945 39.904 506.433 39.064C504.921 38.2 503.733 37 502.869 35.464C502.005 33.904 501.573 32.104 501.573 30.064C501.573 28.048 502.017 26.26 502.905 24.7C503.793 23.14 505.005 21.94 506.541 21.1C508.077 20.26 509.793 19.84 511.689 19.84C513.585 19.84 515.301 20.26 516.837 21.1C518.373 21.94 519.585 23.14 520.473 24.7C521.361 26.26 521.805 28.048 521.805 30.064C521.805 32.08 521.349 33.868 520.437 35.428C519.525 36.988 518.277 38.2 516.693 39.064C515.133 39.904 513.405 40.324 511.509 40.324ZM511.509 36.76C512.565 36.76 513.549 36.508 514.461 36.004C515.397 35.5 516.153 34.744 516.729 33.736C517.305 32.728 517.593 31.504 517.593 30.064C517.593 28.624 517.317 27.412 516.765 26.428C516.213 25.42 515.481 24.664 514.569 24.16C513.657 23.656 512.673 23.404 511.617 23.404C510.561 23.404 509.577 23.656 508.665 24.16C507.777 24.664 507.069 25.42 506.541 26.428C506.013 27.412 505.749 28.624 505.749 30.064C505.749 32.2 506.289 33.856 507.369 35.032C508.473 36.184 509.853 36.76 511.509 36.76ZM530.65 23.512V34.492C530.65 35.236 530.818 35.776 531.154 36.112C531.514 36.424 532.114 36.58 532.954 36.58H535.474V40H532.234C530.386 40 528.97 39.568 527.986 38.704C527.002 37.84 526.51 36.436 526.51 34.492V23.512H524.17V20.164H526.51V15.232H530.65V20.164H535.474V23.512H530.65ZM544.045 23.512V34.492C544.045 35.236 544.213 35.776 544.549 36.112C544.909 36.424 545.509 36.58 546.349 36.58H548.869V40H545.629C543.781 40 542.365 39.568 541.381 38.704C540.397 37.84 539.905 36.436 539.905 34.492V23.512H537.565V20.164H539.905V15.232H544.045V20.164H548.869V23.512H544.045ZM570.507 20.164L558.339 49.324H554.091L558.123 39.676L550.311 20.164H554.883L560.463 35.284L566.259 20.164H570.507ZM577.708 13.36V40H573.604V13.36H577.708ZM581.728 29.992C581.728 28 582.136 26.236 582.952 24.7C583.792 23.164 584.92 21.976 586.336 21.136C587.776 20.272 589.36 19.84 591.088 19.84C592.648 19.84 594.004 20.152 595.156 20.776C596.332 21.376 597.268 22.132 597.964 23.044V20.164H602.104V40H597.964V37.048C597.268 37.984 596.32 38.764 595.12 39.388C593.92 40.012 592.552 40.324 591.016 40.324C589.312 40.324 587.752 39.892 586.336 39.028C584.92 38.14 583.792 36.916 582.952 35.356C582.136 33.772 581.728 31.984 581.728 29.992ZM597.964 30.064C597.964 28.696 597.676 27.508 597.1 26.5C596.548 25.492 595.816 24.724 594.904 24.196C593.992 23.668 593.008 23.404 591.952 23.404C590.896 23.404 589.912 23.668 589 24.196C588.088 24.7 587.344 25.456 586.768 26.464C586.216 27.448 585.94 28.624 585.94 29.992C585.94 31.36 586.216 32.56 586.768 33.592C587.344 34.624 588.088 35.416 589 35.968C589.936 36.496 590.92 36.76 591.952 36.76C593.008 36.76 593.992 36.496 594.904 35.968C595.816 35.44 596.548 34.672 597.1 33.664C597.676 32.632 597.964 31.432 597.964 30.064ZM611.599 23.116C612.295 22.156 613.243 21.376 614.443 20.776C615.667 20.152 617.023 19.84 618.511 19.84C620.263 19.84 621.847 20.26 623.263 21.1C624.679 21.94 625.795 23.14 626.611 24.7C627.427 26.236 627.835 28 627.835 29.992C627.835 31.984 627.427 33.772 626.611 35.356C625.795 36.916 624.667 38.14 623.227 39.028C621.811 39.892 620.239 40.324 618.511 40.324C616.975 40.324 615.607 40.024 614.407 39.424C613.231 38.824 612.295 38.056 611.599 37.12V40H607.495V13.36H611.599V23.116ZM623.659 29.992C623.659 28.624 623.371 27.448 622.795 26.464C622.243 25.456 621.499 24.7 620.563 24.196C619.651 23.668 618.667 23.404 617.611 23.404C616.579 23.404 615.595 23.668 614.659 24.196C613.747 24.724 613.003 25.492 612.427 26.5C611.875 27.508 611.599 28.696 611.599 30.064C611.599 31.432 611.875 32.632 612.427 33.664C613.003 34.672 613.747 35.44 614.659 35.968C615.595 36.496 616.579 36.76 617.611 36.76C618.667 36.76 619.651 36.496 620.563 35.968C621.499 35.416 622.243 34.624 622.795 33.592C623.371 32.56 623.659 31.36 623.659 29.992ZM639.093 40.324C637.533 40.324 636.129 40.048 634.881 39.496C633.657 38.92 632.685 38.152 631.965 37.192C631.245 36.208 630.861 35.116 630.813 33.916H635.061C635.133 34.756 635.529 35.464 636.249 36.04C636.993 36.592 637.917 36.868 639.021 36.868C640.173 36.868 641.061 36.652 641.685 36.22C642.333 35.764 642.657 35.188 642.657 34.492C642.657 33.748 642.297 33.196 641.577 32.836C640.881 32.476 639.765 32.08 638.229 31.648C636.741 31.24 635.529 30.844 634.593 30.46C633.657 30.076 632.841 29.488 632.145 28.696C631.473 27.904 631.137 26.86 631.137 25.564C631.137 24.508 631.449 23.548 632.073 22.684C632.697 21.796 633.585 21.1 634.737 20.596C635.913 20.092 637.257 19.84 638.769 19.84C641.025 19.84 642.837 20.416 644.205 21.568C645.597 22.696 646.341 24.244 646.437 26.212H642.333C642.261 25.324 641.901 24.616 641.253 24.088C640.605 23.56 639.729 23.296 638.625 23.296C637.545 23.296 636.717 23.5 636.141 23.908C635.565 24.316 635.277 24.856 635.277 25.528C635.277 26.056 635.469 26.5 635.853 26.86C636.237 27.22 636.705 27.508 637.257 27.724C637.809 27.916 638.625 28.168 639.705 28.48C641.145 28.864 642.321 29.26 643.233 29.668C644.169 30.052 644.973 30.628 645.645 31.396C646.317 32.164 646.665 33.184 646.689 34.456C646.689 35.584 646.377 36.592 645.753 37.48C645.129 38.368 644.241 39.064 643.089 39.568C641.961 40.072 640.629 40.324 639.093 40.324Z"
-              fill={theme.palette.text.secondary}
+              fill={styles.secondaryTextColor}
             />
             <path
               d="M115.1 59.625V61.95H107.725V67.075H113.475V69.4H107.725V77H104.875V59.625H115.1ZM130.366 69.775C130.366 70.2917 130.333 70.7583 130.266 71.175H119.741C119.825 72.275 120.233 73.1583 120.966 73.825C121.7 74.4917 122.6 74.825 123.666 74.825C125.2 74.825 126.283 74.1833 126.916 72.9H129.991C129.575 74.1667 128.816 75.2083 127.716 76.025C126.633 76.825 125.283 77.225 123.666 77.225C122.35 77.225 121.166 76.9333 120.116 76.35C119.083 75.75 118.266 74.9167 117.666 73.85C117.083 72.7667 116.791 71.5167 116.791 70.1C116.791 68.6833 117.075 67.4417 117.641 66.375C118.225 65.2917 119.033 64.4583 120.066 63.875C121.116 63.2917 122.316 63 123.666 63C124.966 63 126.125 63.2833 127.141 63.85C128.158 64.4167 128.95 65.2167 129.516 66.25C130.083 67.2667 130.366 68.4417 130.366 69.775ZM127.391 68.875C127.375 67.825 127 66.9833 126.266 66.35C125.533 65.7167 124.625 65.4 123.541 65.4C122.558 65.4 121.716 65.7167 121.016 66.35C120.316 66.9667 119.9 67.8083 119.766 68.875H127.391ZM136.021 65.275C136.504 64.6083 137.163 64.0667 137.996 63.65C138.846 63.2167 139.788 63 140.821 63C142.038 63 143.138 63.2917 144.121 63.875C145.104 64.4583 145.879 65.2917 146.446 66.375C147.013 67.4417 147.296 68.6667 147.296 70.05C147.296 71.4333 147.013 72.675 146.446 73.775C145.879 74.8583 145.096 75.7083 144.096 76.325C143.113 76.925 142.021 77.225 140.821 77.225C139.754 77.225 138.804 77.0167 137.971 76.6C137.154 76.1833 136.504 75.65 136.021 75V77H133.171V58.5H136.021V65.275ZM144.396 70.05C144.396 69.1 144.196 68.2833 143.796 67.6C143.413 66.9 142.896 66.375 142.246 66.025C141.613 65.6583 140.929 65.475 140.196 65.475C139.479 65.475 138.796 65.6583 138.146 66.025C137.513 66.3917 136.996 66.925 136.596 67.625C136.213 68.325 136.021 69.15 136.021 70.1C136.021 71.05 136.213 71.8833 136.596 72.6C136.996 73.3 137.513 73.8333 138.146 74.2C138.796 74.5667 139.479 74.75 140.196 74.75C140.929 74.75 141.613 74.5667 142.246 74.2C142.896 73.8167 143.413 73.2667 143.796 72.55C144.196 71.8333 144.396 71 144.396 70.05ZM155.883 73.2V71.05L164.283 59H167.733V70.725H170.058V73.2H167.733V77H164.933V73.2H155.883ZM165.058 61.9L159.158 70.725H165.058V61.9ZM183.37 66.75V69.15H173.095V66.75H183.37ZM196.844 63C196.661 62.2333 196.327 61.6583 195.844 61.275C195.361 60.8917 194.686 60.7 193.819 60.7C192.502 60.7 191.519 61.2083 190.869 62.225C190.219 63.225 189.877 64.8917 189.844 67.225C190.277 66.4917 190.911 65.925 191.744 65.525C192.577 65.125 193.477 64.925 194.444 64.925C195.544 64.925 196.511 65.1667 197.344 65.65C198.194 66.1167 198.852 66.8 199.319 67.7C199.802 68.6 200.044 69.675 200.044 70.925C200.044 72.0917 199.811 73.1333 199.344 74.05C198.894 74.9667 198.219 75.6917 197.319 76.225C196.419 76.7417 195.336 77 194.069 77C192.352 77 191.002 76.625 190.019 75.875C189.052 75.1083 188.369 74.0333 187.969 72.65C187.586 71.2667 187.394 69.5417 187.394 67.475C187.394 61.375 189.544 58.325 193.844 58.325C195.511 58.325 196.819 58.775 197.769 59.675C198.736 60.575 199.302 61.6833 199.469 63H196.844ZM193.819 67.325C192.802 67.325 191.936 67.6333 191.219 68.25C190.519 68.85 190.169 69.725 190.169 70.875C190.169 72.025 190.494 72.9417 191.144 73.625C191.811 74.2917 192.736 74.625 193.919 74.625C194.969 74.625 195.794 74.3 196.394 73.65C197.011 73 197.319 72.1333 197.319 71.05C197.319 69.9167 197.019 69.0167 196.419 68.35C195.836 67.6667 194.969 67.325 193.819 67.325ZM206.584 73.575L203.559 80.475H201.709L203.484 73.575H206.584ZM216.768 73.125C218.368 71.7417 219.627 70.6083 220.543 69.725C221.477 68.825 222.252 67.8917 222.868 66.925C223.485 65.9583 223.793 64.9917 223.793 64.025C223.793 63.025 223.552 62.2417 223.068 61.675C222.602 61.1083 221.86 60.825 220.843 60.825C219.86 60.825 219.093 61.1417 218.543 61.775C218.01 62.3917 217.727 63.225 217.693 64.275H214.943C214.993 62.375 215.56 60.925 216.643 59.925C217.743 58.9083 219.135 58.4 220.818 58.4C222.635 58.4 224.052 58.9 225.068 59.9C226.102 60.9 226.618 62.2333 226.618 63.9C226.618 65.1 226.31 66.2583 225.693 67.375C225.093 68.475 224.368 69.4667 223.518 70.35C222.685 71.2167 221.618 72.225 220.318 73.375L219.193 74.375H227.118V76.75H214.968V74.675L216.768 73.125ZM229.622 67.65C229.622 64.7667 230.105 62.5167 231.072 60.9C232.055 59.2667 233.739 58.45 236.122 58.45C238.505 58.45 240.18 59.2667 241.147 60.9C242.13 62.5167 242.622 64.7667 242.622 67.65C242.622 70.5667 242.13 72.85 241.147 74.5C240.18 76.1333 238.505 76.95 236.122 76.95C233.739 76.95 232.055 76.1333 231.072 74.5C230.105 72.85 229.622 70.5667 229.622 67.65ZM239.822 67.65C239.822 66.3 239.73 65.1583 239.547 64.225C239.38 63.2917 239.03 62.5333 238.497 61.95C237.964 61.35 237.172 61.05 236.122 61.05C235.072 61.05 234.28 61.35 233.747 61.95C233.214 62.5333 232.855 63.2917 232.672 64.225C232.505 65.1583 232.422 66.3 232.422 67.65C232.422 69.05 232.505 70.225 232.672 71.175C232.839 72.125 233.189 72.8917 233.722 73.475C234.272 74.0583 235.072 74.35 236.122 74.35C237.172 74.35 237.964 74.0583 238.497 73.475C239.047 72.8917 239.405 72.125 239.572 71.175C239.739 70.225 239.822 69.05 239.822 67.65ZM247.213 73.125C248.813 71.7417 250.071 70.6083 250.988 69.725C251.921 68.825 252.696 67.8917 253.313 66.925C253.929 65.9583 254.238 64.9917 254.238 64.025C254.238 63.025 253.996 62.2417 253.513 61.675C253.046 61.1083 252.304 60.825 251.288 60.825C250.304 60.825 249.538 61.1417 248.988 61.775C248.454 62.3917 248.171 63.225 248.138 64.275H245.388C245.438 62.375 246.004 60.925 247.088 59.925C248.188 58.9083 249.579 58.4 251.263 58.4C253.079 58.4 254.496 58.9 255.513 59.9C256.546 60.9 257.063 62.2333 257.063 63.9C257.063 65.1 256.754 66.2583 256.138 67.375C255.538 68.475 254.813 69.4667 253.963 70.35C253.129 71.2167 252.063 72.225 250.763 73.375L249.638 74.375H257.563V76.75H245.413V74.675L247.213 73.125ZM261.641 73.125C263.241 71.7417 264.5 70.6083 265.416 69.725C266.35 68.825 267.125 67.8917 267.741 66.925C268.358 65.9583 268.666 64.9917 268.666 64.025C268.666 63.025 268.425 62.2417 267.941 61.675C267.475 61.1083 266.733 60.825 265.716 60.825C264.733 60.825 263.966 61.1417 263.416 61.775C262.883 62.3917 262.6 63.225 262.566 64.275H259.816C259.866 62.375 260.433 60.925 261.516 59.925C262.616 58.9083 264.008 58.4 265.691 58.4C267.508 58.4 268.925 58.9 269.941 59.9C270.975 60.9 271.491 62.2333 271.491 63.9C271.491 65.1 271.183 66.2583 270.566 67.375C269.966 68.475 269.241 69.4667 268.391 70.35C267.558 71.2167 266.491 72.225 265.191 73.375L264.066 74.375H271.991V76.75H259.841V74.675L261.641 73.125Z"
-              fill={theme.palette.text.secondary}
+              fill={styles.secondaryTextColor}
             />
             <TartanHacksIcon
               width="75"
