+++ src/components/design/WaveBackground/index.module.scss
@@ -1,0 +1,45 @@
+@import "~/styles/variables.scss";
+@import "~/styles/mixins.scss";
+
+.waveContainer {
+  position: fixed;
+  width: 100%;
+  overflow: hidden;
+  line-height: 0;
+  z-index: -1;
+}
+
+.waveSvg {
+  position: absolute;
+  width: 100%;
+  height: 465px;
+  bottom: 0;
+  filter: drop-shadow(0px 4px 8px rgba(0, 0, 0, 0.2));
+
+  @include down($breakpoint-md){
+    height: 15em; 
+  }
+
+  @include down($breakpoint-sm){
+    height: 10em; 
+  }
+}
+
+.scottyIcon {
+  position: relative;
+  width: 75px;
+  height: 65px;
+
+  @include down($breakpoint-md){
+    width: 100%; 
+  }
+
+  margin-right: 1em; 
+}
+
+.gradientStart {
+  stop-color: $wave-gradient-color-start; 
+}
+.gradientStop {
+  stop-color: $wave-gradient-color-end;
+}
