+++ styles/Confirmation.module.scss
@@ -1,0 +1,10 @@
+@import "./variables.scss";
+@import "./mixins.scss";
+
+.dialog {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
